% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{BufferCrop}
\alias{BufferCrop}
\title{Buffer-Crop-Function}
\usage{
BufferCrop(pts, dem, extent)
}
\arguments{
\item{pts}{List of Points (Sites)}

\item{dem}{DEM}

\item{extent}{extent of buffer that will be croped in KM}
}
\value{
dem croped DEM
}
\description{
Buffer-Crop-Function
}
