% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteExploitationTerritories.R
\name{convertDFtoSPDF}
\alias{convertDFtoSPDF}
\title{converting input df of points into SpatialPointsDataFrame()}
\usage{
convertDFtoSPDF(pts)
}
\description{
converting input df of points into SpatialPointsDataFrame()
}
\keyword{internal}
