% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts.R
\docType{data}
\name{pts}
\alias{pts}
\title{Points}
\format{A data frame with 5 rows and 3 variables:
\describe{
  \item{id}{Site ID}
  \item{X}{Longitude in decimal degree}
  \item{Y}{Latitide in decimal degree}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
pts
}
\description{
A dataset containing five sites with their respective coordinates.
}
\keyword{datasets}
