% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteExploitationTerritories.R
\name{SiteExplorationTerritories}
\alias{SiteExplorationTerritories}
\title{Site Exploration Territories}
\usage{
SiteExplorationTerritories(pts, dem, slope, epsg, TheoreticalSlopes = seq(-70,
  70, 1), numberOfNeighbors = 8, damping = FALSE, dampingFactor = 16,
  numberOfDirections = 8, timeOfInterest = 2, numberOfIsochrones = 2,
  intervallOfIsochrones = 1)
}
\arguments{
\item{pts}{Points}

\item{dem}{DEM}

\item{epsg}{EPSG}

\item{TheoreticalSlopes}{a sequence}

\item{numberOfNeighbors}{an integer, only if no SLOPE input raster available, select algorithm to calculate slope: 4 == 4 neighbours, smooth surfaces, Fleming and Hoffer (1979), Ritter (1987) (see raster package terrain); 8 == 8 neighbours, rough surfaces, Horn (1981) (see raster package terrain)}

\item{damping}{TRUE/FALSE; damp hiking speed if slope is steeper than the dampingFactor (slope in degrees)}

\item{dampingFactor}{Damping Factor}

\item{numberOfDirections}{Number of Directions; move cases for transition layer; select move case for calculating time-cost; cell connections, directions for transition: 4 == 4 directions, Rook case; 8 == 8 directions, Queen case; 16 == 16 directions, knight move}

\item{timeOfInterest}{Time interval of interest; select a time interval of interest; use integer > 0}

\item{numberOfIsochrones}{Select a number and interval of isochrones output; number of isochrones [hours], integer}

\item{intervallOfIsochrones}{Intervals of insochrones [hours], numeric}

\item{dem}{slope raster}
}
\value{
returns a list containing for each Site a list of the Accumulated Cost Surface and Isochrones; call by using variable[[a]][[b]], while a == 1 for the first site and b == 1 for rAccumulatedCostSurface.h and b == 2 for vContourLines
}
\description{
based on Technical Note 3: Distance relationships or does distance matter – calculating a non-isotropic  spatial  relationship  by  integrating  human  energy  expenditure  in  terrain based estimations – A seamless workflow for defining archaeological Site Exploitation Territories (SET) using the open source (geo-)statistical language R (Ahlrichs, Jan Johannes, Gries, Philipp and Schmidt, Karsten) see: http://www.uni-tuebingen.de/index.php?eID=tx_nawsecuredl&u=0&g=0&t=1499250113&hash=af99ce857ef61c6df5a2b70a907cf91473fa6deb&file=fileadmin/Uni_Tuebingen/SFB/SFB_1070/dokumente/technical_notes/technical_notes_3/CRC1070_2016_Technical_Note_03_SET.pdf
}
