% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiteExploitationTerritories.R
\name{ToblersHikingFunction}
\alias{ToblersHikingFunction}
\title{implementing tobler's hiking function}
\usage{
ToblersHikingFunction(x)
}
\description{
implementing tobler's hiking function
}
\keyword{internal}
