% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R, R/myfuntions.R
\name{trans}
\alias{trans}
\alias{trans}
\title{Transformierung eines Layers in die gewünschte Projektion}
\usage{
trans(layer, EPSG)

trans(layer, EPSG)
}
\description{
Transformierung eines Layers in die gewünschte Projektion

Transformierung eines Layers in die gewünschte Projektion
}
\keyword{internal}
